mode search models into separate volume
