+++ .gitignore
@@ -130,3 +130,4 @@ dmypy.json
 
 telegram-token.txt
 mongo-data/
+search-data/
