--- search/data.dvc
@@ -1,5 +1,0 @@
-outs:
-- md5: 231ae7bbee3d41f2b0235696e182efea.dir
-  size: 2917081441
-  nfiles: 11
-  path: data
