+++ search/src/engine/IndexSearch.py
@@ -1,14 +1,8 @@
 import scann
-import logging
-from config import settings
-
-filename = settings['INDEX_FILENAME']
-nprobe = settings['INDEX_NPROBE']
 
 class IndexSearch:
 
-    def __init__(self):
-        logging.info(f'Reading index from {filename}')
+    def __init__(self, filename):
         self.index = scann.scann_ops_pybind.load_searcher(filename)
 
     def search(self, questions_embeding, passages_count_to_search):
