+++ search/src/engine/__init__.py
@@ -1,5 +1,3 @@
 from .init_engine import load_and_init_engine
 from .SearchEngine import SearchEngine
 from .EmbedingModel import EmbedingModel
-from .read_models import read_models
-from .PassagesDatabase import PassagesInMemoryDatabase
